- Restored serialization extension tests
