+++ src/MooVC.Tests/Serialization/ObjectExtensions.Clone.cs
@@ -1,14 +1,12 @@
 ﻿namespace MooVC.Serialization
 {
-    using System.Collections.Generic;
-
     internal static partial class ObjectExtensions
     {
         public static T Clone<T>(this T original)
         {
-            IEnumerable<byte> data = original.Serialize();
-
-            return data.Deserialize<T>();
+            return original
+                .Serialize()
+                .Deserialize<T>();
         }
     }
 }
