+++ src/MooVC.Tests/Serialization/StringExtensions.Deserialize.cs
@@ -1,0 +1,12 @@
+﻿namespace MooVC.Serialization
+{
+    using Newtonsoft.Json;
+
+    internal static partial class EnumerableExtensions
+    {
+        public static T Deserialize<T>(this string json)
+        {
+            return JsonConvert.DeserializeObject<T>(json, Settings.Default)!;
+        }
+    }
+}
